added frontend of portfolio

+++ backend/portfolio/models.py
@@ -6,7 +6,7 @@ from stocks.models import Stock
 class stockHolding(models.Model):
     user = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
     stock = models.ForeignKey(Stock, on_delete=models.CASCADE)
-    quantity = models.PositiveIntegerField
+    quantity = models.DecimalField(max_digits=10, decimal_places=2)
     average_buy_price = models.DecimalField(max_digits=12, decimal_places=2)
     added_on = models.DateTimeField(auto_now_add=True)
